Updates and improvements before deployment to other chains (#238)
* QoL improvement for repaying max within Borrower

* Improve ERC4626 name by removing unicode characters

* Allow anyone to claim rewards on behalf of anyone else

* Increase oracle cardinality safety factor

* Remove unnecessary cast in VolatilityOracle

* Stronger protection on Lender shares conversion rate

+++ core/src/libraries/constants/Constants.sol
@@ -159,7 +159,7 @@ uint256 constant FEE_GROWTH_AVG_WINDOW = 72 hours;
 /// @dev The length of the circular buffer that stores feeGrowthGlobals samples.
 /// Must be in interval
 /// \\( \left[ \frac{\text{FEE_GROWTH_AVG_WINDOW}}{\text{FEE_GROWTH_SAMPLE_PERIOD}}, 256 \right) \\)
-uint256 constant FEE_GROWTH_ARRAY_LENGTH = 32;
+uint8 constant FEE_GROWTH_ARRAY_LENGTH = 32;
 
 /// @dev The minimum number of seconds that must elapse before a new feeGrowthGlobals sample will be stored. This
 /// controls how often the oracle can update IV.
